Update
- Optimzie `List<>.MoveTo`
- Remove duplicate cast in `ReversedSpan`
- Add more API to  `List<>.SortedModifier`
- Add accessor extensions for `Span<>`
- Optimize `Result` creation apis
- Fix bug on `CatchCancellation` (well because i didnt test
- Add `TraThrow.ThrowSwitchExpressionException`

+++ src/Trarizon.Library/Collections/StackAlloc/ReadOnlyReversedSpan.cs
@@ -31,7 +31,7 @@ public readonly ref struct ReadOnlyReversedSpan<T>
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public ref readonly T DangerousGetReferenceAt(int index)
-        => ref Unsafe.Subtract(ref Unsafe.AsRef(in _reference), (nint)(uint)index);
+        => ref Unsafe.Subtract(ref Unsafe.AsRef(in _reference), index);
 
 
     public int Length => _length;
